pyproject.toml
[tool.poetry]
name = "pizza-bot"
version = "0.1.0"
description = ""
authors = ["balancy <balancy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
python-dotenv = "^0.19.2"
python-slugify = "^5.0.2"
python-telegram-bot = "^13.8.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.11b1", allow-prereleases = true}
ipython = "^7.30.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
